requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
django-widget-tweaks>=1.4.12
reportlab>=4.0.4
django-chartjs>=2.3.0
django-filter>=23.2
django-notifications-hq>=1.8.0